backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
sqlmodel
pydantic
